workers/requirements.txt
# ============================================================
# CYBERDUDEBIVASH AI SYSTEM — PRODUCTION REQUIREMENTS (PINNED)
# All versions pinned for reproducible builds
# ============================================================

# ── Core Framework ────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
starlette==0.41.3

# ── AI / LLM ─────────────────────────────────────────────────
openai==1.51.0
httpx==0.27.2

# ── Task Queue ───────────────────────────────────────────────
celery[redis]==5.4.0
redis==5.1.0
kombu==5.3.7

# ── Database ─────────────────────────────────────────────────
sqlalchemy==2.0.35
alembic==1.13.3

# ── Auth / Security ───────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
cryptography==43.0.3

# ── Validation ───────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# ── Config ───────────────────────────────────────────────────
python-dotenv==1.0.1
pyyaml==6.0.2

# ── Monitoring ───────────────────────────────────────────────
prometheus-client==0.21.0

# ── Utilities ────────────────────────────────────────────────
python-dateutil==2.9.0
requests==2.32.3
